Rework most of .contramap usages. Promap is now broken
