+++ src/helpers/getScaleStyle.js
@@ -1,0 +1,16 @@
+const calculateTransform = oldTransform => scaling =>
+  oldTransform ? `${oldTransform} scale(${scaling})` : `scale(${scaling})`
+
+// getScaleStyle : (Props -> Number) -> Component -> Component
+export default getScaleFromProps =>
+  props => ({
+    ...props,
+    style: {
+      ...props.style,
+      transform: calculateTransform(
+        props && props.style && props.style.transform
+      )(
+        getScaleFromProps(props)
+      ),
+    },
+  })
